gui.edit.htmleditor: fix a warning
git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@5731 3fb6f576-06d4-0310-ad11-a4b695aa9f05

+++ lib/advene/gui/edit/htmleditor.py
@@ -192,7 +192,11 @@ class HTMLEditor(textview_class, HTMLParser):
                                                  b.get_iter_at_mark(m),
                                                  b.get_iter_at_mark(m._endmark))
                             b.delete_mark(m)
-                            b.delete_mark(m._endmark)
+                            if not m._endmark.get_deleted():
+                                # Could already be deleted (for
+                                # instance, br tag have startmark ==
+                                # endmark)
+                                b.delete_mark(m._endmark)
                         except AttributeError, e:
                             print "Exception for %s" % m._tag, unicode(e).encode('utf-8')
                     elif  hasattr(m, '_endtag'):
